+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsejsonschema.py
@@ -184,9 +184,9 @@ class EvaluatorResponseJSONSchema(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-30T06:05:21.250Z"
+    created: Optional[str] = "2026-06-30T07:21:18.786Z"
 
-    updated: Optional[str] = "2026-06-30T06:05:21.250Z"
+    updated: Optional[str] = "2026-06-30T07:21:18.786Z"
 
     updated_by_id: OptionalNullable[str] = UNSET