+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWBJ4806AVJB5BZBGHKZ9YJJ"
+        "01KWBPFPBTANRG7ZVGR0Y3NCBF"
     )
     r"""The unique identifier of the data source"""