+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWBJ480A92Z2QD2NKGWNZV83"
+        "01KWBPFPBYGZRRZVS74DHJ16A4"
     )
     r"""The unique identifier of the data source"""