+++ packages/orq-rc/RELEASES.md
@@ -9089,3 +9089,13 @@ Based on:
 - [python v4.11.0-rc.58] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.58] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.58 - packages/orq-rc
+
+## 2026-06-30 07:24:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.59] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.59] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.59 - packages/orq-rc